crypto: rename threshold to eddy
We decided to call the specific instantiation of flow encryption "eddy", to
distinguish the specific construction from the ideal functionality.
