+++ crypto/src/eddy.rs
