--- crypto/src/threshold.rs
@@ -1,1 +1,0 @@
-pub mod threshold;
